+++ NEWS.md
@@ -2,6 +2,7 @@
 * Add `fish` and `fish_raw` data.
 * Add `summary` method for `occumbData` class.
 * Add `summary` method for `occumbFit` class.
+* Add `occumbGof` class and methods.
 * Change defaults for `cores` arguments for `gof()`, `eval_util_L()`, and `eval_util_L()` functions.
 * Fix `eval_util_L()` and `eval_util_R()` examples.
 * Fix a bug in `add_attribute4()` function.